+++ src/backend/people/api_urls.py
@@ -5,27 +5,30 @@ from django.urls import include, path, re_path
 
 from rest_framework.routers import DefaultRouter
 
-from core.api import viewsets
+from core.api.viewsets.config import ConfigView
+from core.api.viewsets.contacts import ContactViewSet
+from core.api.viewsets.teams import InvitationViewset, TeamAccessViewSet, TeamViewSet
+from core.api.viewsets.users import UserViewSet
 from core.authentication.urls import urlpatterns as oidc_urls
 from core.resource_server.urls import urlpatterns as resource_server_urls
 
 # - Main endpoints
 router = DefaultRouter()
-router.register("contacts", viewsets.ContactViewSet, basename="contacts")
-router.register("teams", viewsets.TeamViewSet, basename="teams")
-router.register("users", viewsets.UserViewSet, basename="users")
+router.register("contacts", ContactViewSet, basename="contacts")
+router.register("teams", TeamViewSet, basename="teams")
+router.register("users", UserViewSet, basename="users")
 
 # - Routes nested under a team
 team_related_router = DefaultRouter()
 team_related_router.register(
     "accesses",
-    viewsets.TeamAccessViewSet,
+    TeamAccessViewSet,
     basename="team_accesses",
 )
 
 team_related_router.register(
     "invitations",
-    viewsets.InvitationViewset,
+    InvitationViewset,
     basename="invitations",
 )
 
@@ -46,5 +49,5 @@ urlpatterns = [
         ),
     ),
     path(f"api/{settings.API_VERSION}/", include("mailbox_manager.urls")),
-    path(f"api/{settings.API_VERSION}/config/", viewsets.ConfigView.as_view()),
+    path(f"api/{settings.API_VERSION}/config/", ConfigView.as_view()),
 ]